requirements.txt
pillow==10.4.0
opencv-python==4.11.0.86
#opencv-contrib-python==4.10.0.84
